+++ src/ui/world_view.h
@@ -46,7 +46,7 @@ class WorldView : public QGraphicsView {
     void keyPressEvent(QKeyEvent* event) override;
 
  public slots:
-    void change_scale(float scale_factor);
+    void change_scale(double scale_factor);
     void reset_scale();
     void scale_to_see_all_items();
 
@@ -57,7 +57,7 @@ class WorldView : public QGraphicsView {
  protected:
  private:
     QBrush background{};
-    float xy_scale{1.0};
+    double xy_scale{1.0};
 };
 
 #endif  // _UI_WORLD_VIEW_H_
